refactor(datasets): seal `datasets` behind api::datasets; remove unmasked dead code
`pub mod datasets` was an unsealed dual-path: the wheel reached the raw
`kglite_core::datasets::*` while the C ABI used the curated
`api::datasets::*`, and the chokepoint gate only guarded `graph::`. Close
the gap the same way `graph` was sealed.

- core: `pub mod datasets` -> `pub(crate) mod datasets`; the `api::datasets`
  re-exports still resolve (facade pattern). Extend
  scripts/check_api_chokepoint.sh to count `(graph|datasets)::` reaches —
  all wrappers at 0.
- wheel: migrate the dataset pyo3 layer (sec/sodir/wikidata.rs) from raw
  `kglite_core::datasets::*` onto `kglite_core::api::datasets::*`.
- dead code unmasked by the seal (pub items in a pub module aren't
  dead-flagged; pub(crate) ones are) — removed: the wholly-dead master-idx
  parser `parsers/idx.rs` (production-unused), ~13 redundant `pub use`
  re-exports of internal parser types, `daily_company_idx_url`,
  `rate_limit_cost_seconds`, `bool_str`, `opt_int`, `mark_company_seen`,
  `METADATA_URL`, and the never-read `RemoteMeta::content_length` field
  (+ the now-pointless unit tests of the removed fns).
- integration tests reaching internals: `datasets_sec_fetch_live` migrated
  to `api::datasets` (dropped its dead-parser assertion); the sodir live
  probe moved in-crate (`sodir/fetch.rs` tests) since it exercises the
  `fetch` submodule; `datasets_sec_idx_parser` (tested only the dead
  parser) deleted.

No behaviour change for any binding — the curated dataset surface is
unchanged; this seals the path and deletes genuinely-dead code. Gate green
(lint, dataset unit + Python tests). Network integration tests are
env-gated and compile-verified (not run locally).
