+++ crates/kglite/src/datasets/sodir/fetch.rs
@@ -202,6 +202,42 @@ mod tests {
     use super::*;
     use serde_json::json;
 
+    /// Live integration test against the Sodir FactMaps REST API.
+    /// Skipped unless `SODIR_LIVE_TEST` is set — CI / offline runs must not
+    /// depend on the network. Lives in-crate (rather than `tests/`) so it can
+    /// reach the `fetch` submodule after `datasets` was sealed to `pub(crate)`.
+    #[tokio::test]
+    async fn fetch_quadrant_live() {
+        if std::env::var("SODIR_LIVE_TEST").is_err() {
+            eprintln!("skipping fetch_quadrant_live — set SODIR_LIVE_TEST=1 to run");
+            return;
+        }
+        let client = ArcGISClient::new().expect("client constructs");
+        // `quadrant` is one of the smallest datasets — good for a probe.
+        let n = count(&client, "quadrant")
+            .await
+            .expect("count probe succeeds");
+        assert!(n > 0, "quadrant should report a non-zero row count");
+
+        let tmp = tempfile::tempdir().unwrap();
+        let csv_path = tmp.path().join("quadrant.csv");
+        let written = fetch_to_csv(&client, "quadrant", &csv_path)
+            .await
+            .expect("fetch_to_csv succeeds");
+        assert!(written > 0, "expected rows written");
+        assert!(csv_path.is_file(), "csv file should exist");
+
+        let content = std::fs::read_to_string(&csv_path).unwrap();
+        assert!(
+            content.lines().count() > 1,
+            "csv should have a header + data rows"
+        );
+        assert!(
+            content.lines().next().unwrap().contains("wkt_geometry"),
+            "geometry layer header should include wkt_geometry"
+        );
+    }
+
     #[test]
     fn cell_rendering() {
         assert_eq!(cell_to_string(None), "");
