+++ crates/kglite/src/datasets/sodir/mod.rs
@@ -32,11 +32,8 @@ pub mod layout;
 pub mod orchestrator;
 pub mod preprocess;
 
-pub use blueprint::{datasets_used_by_blueprint, deep_merge, merge_blueprint_json};
-pub use catalog::{is_known, kind_of, resolve, DataKind};
+pub use blueprint::{datasets_used_by_blueprint, merge_blueprint_json};
 pub use client::ArcGISClient;
 pub use error::{Result, SodirError};
-pub use index::{Action, DatasetEntry, SodirIndex};
 pub use layout::{StorageMode, Workdir};
-pub use orchestrator::{fetch_all, refresh_csvs, FetchAllReport, RefreshReport};
-pub use preprocess::{apply as apply_preprocess, PreprocessReport};
+pub use orchestrator::{fetch_all, FetchAllReport};
